+++ app/page.tsx
@@ -1,11 +1,11 @@
 import { Footer } from '#/ui/Footer';
-import { Header } from '#/ui/Header';
+import { HomeHeader } from '#/ui/HomeHeader';
 import { HomePage } from '#/ui/HomePage';
 
 export default async function Page() {
   return (
     <div>
-      <Header />
+      <HomeHeader />
 
       <HomePage />
