+++ app/projects/[slug]/page.tsx
@@ -2,13 +2,13 @@ import { kebabCase } from 'lodash';
 
 import { fetchData } from '#/lib/api.js';
 import { Footer } from '#/ui/Footer';
-import { Header } from '#/ui/Header';
 import { ProjectPage } from '#/ui/ProjectPage';
+import { Header } from '#/ui/Header';
 
-export default async function Page({ params }) {
+export default function Page({ params }) {
   return (
     <div>
-      <Header />
+      <Header showAboutButton={true} />
 
       <ProjectPage projectSlug={params.slug} />
